Rename theme config
