+++ src/components/LeftRightChart/Axis.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import theme from '../victorytheme';
+import theme from '../../theme';
 
 const TSpan = styled.tspan`
   font-size: 14px;
